+++ Source/SwiftLintBuiltInRules/Rules/Style/SortedImportsRule.swift
@@ -1,59 +1,7 @@
-import Foundation
-import SourceKittenFramework
+import SwiftSyntax
 
-fileprivate extension Line {
-    var contentRange: NSRange {
-        NSRange(location: range.location, length: content.bridge().length)
-    }
-
-    // `Line` in this rule always contains word import
-    // This method returns contents of line that are before import
-    func importAttributes() -> String {
-        content[importAttributesRange()].trimmingCharacters(in: .whitespaces)
-    }
-
-    // `Line` in this rule always contains word import
-    // This method returns contents of line that are after import
-    func importModule() -> Substring {
-        content[importModuleRange()]
-    }
-
-    func importAttributesRange() -> Range<String.Index> {
-        let rangeOfImport = content.range(of: "import")
-        precondition(rangeOfImport != nil)
-        return content.startIndex..<rangeOfImport!.lowerBound
-    }
-
-    func importModuleRange() -> Range<String.Index> {
-        let rangeOfImport = content.range(of: "import")
-        precondition(rangeOfImport != nil)
-        let moduleStart = content.rangeOfCharacter(from: CharacterSet.whitespaces.inverted, options: [],
-                                                   range: rangeOfImport!.upperBound..<content.endIndex)
-        return moduleStart!.lowerBound..<content.endIndex
-    }
-}
-
-private extension Sequence where Element == Line {
-    // Groups lines, so that lines that are one after the other
-    // will end up in same group.
-    func grouped() -> [[Line]] {
-        reduce(into: [[]]) { result, line in
-            guard let last = result.last?.last else {
-                result = [[line]]
-                return
-            }
-
-            if last.index == line.index - 1 {
-                result[result.count - 1].append(line)
-            } else {
-                result.append([line])
-            }
-        }
-    }
-}
-
-@DisabledWithoutSourceKit
-struct SortedImportsRule: CorrectableRule, OptInRule {
+@SwiftSyntaxRule(explicitRewriter: true, optIn: true)
+struct SortedImportsRule: Rule {
     var configuration = SortedImportsConfiguration()
 
     static let description = RuleDescription(
@@ -65,86 +13,189 @@ struct SortedImportsRule: CorrectableRule, OptInRule {
         triggeringExamples: SortedImportsRuleExamples.triggeringExamples,
         corrections: SortedImportsRuleExamples.corrections
     )
+}
 
-    func validate(file: SwiftLintFile) -> [StyleViolation] {
-        let groups = importGroups(in: file, filterEnabled: false)
-        return violatingOffsets(inGroups: groups).map { index -> StyleViolation in
-            let location = Location(file: file, characterOffset: index)
-            return StyleViolation(ruleDescription: Self.description,
-                                  severity: configuration.severity,
-                                  location: location)
-        }
-    }
+private extension SortedImportsRule {
+    final class Visitor: ViolationsSyntaxVisitor<ConfigurationType> {
+        private var imports = [Import]()
 
-    private func importGroups(in file: SwiftLintFile, filterEnabled: Bool) -> [[Line]] {
-        var importRanges = file.match(pattern: "import\\s+\\w+", with: [.keyword, .identifier])
-        if filterEnabled {
-            importRanges = file.ruleEnabled(violatingRanges: importRanges, for: self)
+        override var skippableDeclarations: [any DeclSyntaxProtocol.Type] { .all }
+
+        override func visitPost(_ node: ImportDeclSyntax) {
+            imports.append(
+                Import.from(importDecl: node, grouping: configuration.grouping, locationConverter: locationConverter)
+            )
         }
 
-        let contents = file.stringView
-        let lines = file.lines
-        let importLines: [Line] = importRanges.compactMap { range in
-            guard let line = contents.lineAndCharacter(forCharacterOffset: range.location)?.line else {
-                return nil
-            }
-            return lines[line - 1]
-        }
-
-        return importLines.grouped()
-    }
-
-    private func violatingOffsets(inGroups groups: [[Line]]) -> [Int] {
-        groups.flatMap { group in
-            zip(group, group.dropFirst()).reduce(into: []) { violatingOffsets, groupPair in
-                let (previous, current) = groupPair
-                let isOrderedCorrectly = should(previous, comeBefore: current)
-                if isOrderedCorrectly {
-                    return
+        override func visitPost(_: SourceFileSyntax) {
+            // Group imports that are adjacent to each other.
+            var importBlocks = [[Import]]()
+            for `import` in imports {
+                if let lastBlock = importBlocks.last, let lastImport = lastBlock.last {
+                    if `import`.isDirectlyAfter(other: lastImport, in: file) {
+                        importBlocks[importBlocks.count - 1].append(`import`)
+                    } else {
+                        importBlocks.append([`import`])
+                    }
+                } else {
+                    importBlocks.append([`import`])
                 }
-                let distance = current.content.distance(from: current.content.startIndex,
-                                                        to: current.importModuleRange().lowerBound)
-                violatingOffsets.append(current.range.location + distance)
             }
-        }
-    }
 
-    /// - returns: whether `lhs` should come before `rhs` based on a comparison of the contents of the import lines
-    private func should(_ lhs: Line, comeBefore rhs: Line) -> Bool {
-        switch configuration.grouping {
-        case .attributes:
-            let lhsAttributes = lhs.importAttributes()
-            let rhsAttributes = rhs.importAttributes()
-            if lhsAttributes != rhsAttributes {
-                if lhsAttributes.isEmpty != rhsAttributes.isEmpty {
-                    return rhsAttributes.isEmpty
+            // For every block, check that the imports are sorted.
+            for block in importBlocks {
+                for (previous, current) in zip(block, block.dropFirst()) where previous > current {
+                    violations.append(current.violationPosition)
                 }
-                return lhsAttributes < rhsAttributes
             }
-        case .names:
-            break
         }
-        return lhs.importModule().lowercased() <= rhs.importModule().lowercased()
     }
 
-    func correct(file: SwiftLintFile) -> Int {
-        let groups = importGroups(in: file, filterEnabled: true)
-        let offsets = violatingOffsets(inGroups: groups)
-        guard offsets.isNotEmpty else {
-            return 0
-        }
-        let correctedContents = NSMutableString(string: file.contents)
-        for group in groups.map({ $0.sorted(by: should(_:comeBefore:)) }) {
-            guard let first = group.first?.contentRange else {
-                continue
+    final class Rewriter: ViolationsSyntaxRewriter<ConfigurationType> {
+        override func visit(_ node: SourceFileSyntax) -> SourceFileSyntax {
+            var statements = rewrite(statements: node.statements)
+            if numberOfCorrections == 0 {
+                return super.visit(node)
             }
-            let result = group.map(\.content).joined(separator: "\n")
-            let union = group.dropFirst().reduce(first) { result, line in
-                NSUnionRange(result, line.contentRange)
+            if let leadingTrivia = statements.first?.leadingTrivia {
+                statements[0] = statements[0].with(\.leadingTrivia, leadingTrivia.skippingLeadingNewline)
             }
-            correctedContents.replaceCharacters(in: union, with: result)
+            return super.visit(node.with(\.statements, CodeBlockItemListSyntax(statements)))
         }
-        file.write(correctedContents.bridge())
-        return offsets.count
+
+        private func rewrite(statements: CodeBlockItemListSyntax) -> [CodeBlockItemSyntax] {
+            var rewrittenStatements = [CodeBlockItemSyntax]()
+            var imports = [Import]()
+
+            for stmt in statements {
+                if let importDecl = stmt.item.as(ImportDeclSyntax.self) {
+                    if importDecl.isContainedIn(regions: disabledRegions, locationConverter: locationConverter) {
+                        rewrittenStatements.append(stmt)
+                        continue
+                    }
+                    let `import` = Import.from(
+                        importDecl: importDecl,
+                        grouping: configuration.grouping,
+                        locationConverter: locationConverter
+                    )
+                    if let lastImport = imports.last, !`import`.isDirectlyAfter(other: lastImport, in: file) {
+                        rewrittenStatements.append(contentsOf: sort(imports))
+                        imports = [`import`]
+                        continue
+                    }
+                    imports.append(`import`)
+                } else if let ifConfigDecl = stmt.item.as(IfConfigDeclSyntax.self) {
+                    // Recursively rewrite imports inside `#if` blocks.
+                    let rewrittenClauses = ifConfigDecl.clauses.map { clause in
+                        let rewrittenClause = clause.elements.map { elements in
+                            let rewrittenElements = rewrite(statements: elements.as(CodeBlockItemListSyntax.self) ?? [])
+                            return CodeBlockItemListSyntax(rewrittenElements)
+                        }
+                        return clause.with(\.elements, .statements(rewrittenClause ?? []))
+                    }
+                    let rewrittenIfConfig = ifConfigDecl.with(\.clauses, IfConfigClauseListSyntax(rewrittenClauses))
+                    if !imports.isEmpty {
+                        rewrittenStatements.append(contentsOf: sort(imports))
+                        imports = []
+                    }
+                    rewrittenStatements.append(CodeBlockItemSyntax(item: .decl(DeclSyntax(rewrittenIfConfig))))
+                } else {
+                    rewrittenStatements.append(contentsOf: sort(imports))
+                    imports = []
+                    rewrittenStatements.append(stmt)
+                }
+            }
+            rewrittenStatements.append(contentsOf: sort(imports))
+            return rewrittenStatements
+        }
+
+        private func sort(_ imports: [Import]) -> [CodeBlockItemSyntax] {
+            let sorted = imports.sorted(by: { $0 < $1 })
+            numberOfCorrections += imports.difference(from: sorted).count
+            return sorted.map(\.importDecl.asCodeBlockItem)
+        }
+    }
+}
+
+private extension Trivia {
+    var skippingLeadingNewline: Self {
+        if pieces.onlyElement?.isNewline == true {
+            Trivia(pieces: dropLast())
+        } else if containsComments, pieces.first?.isNewline == true {
+            Trivia(pieces: dropFirst())
+        } else {
+            self
+        }
+    }
+}
+private extension ImportDeclSyntax {
+    var asCodeBlockItem: CodeBlockItemSyntax {
+        let item = CodeBlockItemSyntax(item: .decl(DeclSyntax(self)))
+        return leadingTrivia.pieces.first?.isNewline == true
+            ? item
+            : item.with(\.leadingTrivia, [.newlines(1)] + leadingTrivia)
+    }
+}
+
+private struct Import: Comparable {
+    let importDecl: ImportDeclSyntax
+    let line: Int
+    let offset: Int
+    let attributes: String
+
+    var violationPosition: AbsolutePosition {
+        importDecl.path.positionAfterSkippingLeadingTrivia
+    }
+
+    var symbol: String {
+        importDecl.path.map(\.name.text).joined(separator: ".")
+    }
+
+    static func from(importDecl: ImportDeclSyntax,
+                     grouping: SortedImportsConfiguration.Grouping,
+                     locationConverter: SourceLocationConverter) -> Self {
+        let attributes: [String] =
+            if grouping == .attributes {
+                importDecl.attributes.compactMap {
+                    $0.as(AttributeSyntax.self)?.attributeName.as(IdentifierTypeSyntax.self)?.name.text
+                }
+            } else {
+                []
+            }
+        let startPosition =
+            if importDecl.leadingTrivia.containsComments {
+                importDecl.position
+            } else {
+                importDecl.positionAfterSkippingLeadingTrivia
+            }
+        let startLine = locationConverter.location(for: startPosition).line
+        return Self(
+            importDecl: importDecl,
+            line: startLine,
+            offset: locationConverter.location(for: importDecl.path.endPositionBeforeTrailingTrivia).line - startLine,
+            attributes: attributes.joined(),
+        )
+    }
+
+    func isDirectlyAfter(other: Self, in file: SwiftLintFile) -> Bool {
+        let lineAfterOther = other.line + other.offset
+        if line == lineAfterOther {
+            return true
+        }
+        // Check for empty lines or comment-only lines between imports.
+        return !(lineAfterOther..<line).contains { line in
+            let content = file.lines[line - 1].content
+            return content.isEmpty || content.starts(with: "#")
+        }
+    }
+
+    static func < (lhs: Self, rhs: Self) -> Bool {
+        if lhs.attributes != rhs.attributes {
+            if lhs.attributes.isEmpty {
+                return false
+            }
+            return rhs.attributes.isEmpty || lhs.attributes < rhs.attributes
+        }
+        return lhs.symbol.caseInsensitiveCompare(rhs.symbol) == .orderedAscending
     }
 }